# src/main/java/org/drad/movie_tickets/model/Customer.java
package org.drad.movie_tickets.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import javax.validation.constraints.Max;
import javax.validation.constraints.Min;
import javax.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

/**
 * Customer
 */
@AllArgsConstructor
@NoArgsConstructor
@Builder
@EqualsAndHashCode
@Getter
@Setter
public class Customer {
    @JsonProperty("name")
    private String name;

    @JsonProperty("age")
    @Min(0)
    @Max(120)
    @NotNull
    private Integer age;

}